batched-secret-rotation-lambdas/lt-batched-integration-rotation/src/LTDW-Secrets-Rotation/requirements.txt
requests==2.32.5
urllib3==2.6.2
psycopg2-binary==2.9.11